shared token cache base deserves its own module
