+++ sdk/identity/azure-identity/tests/test_shared_cache_credential_async.py
@@ -7,7 +7,7 @@ from unittest.mock import Mock
 from azure.core.exceptions import ClientAuthenticationError
 from azure.identity import KnownAuthorities
 from azure.identity.aio import SharedTokenCacheCredential
-from azure.identity._credentials.shared_cache import (
+from azure.identity._internal.shared_token_cache import (
     MULTIPLE_ACCOUNTS,
     MULTIPLE_MATCHING_ACCOUNTS,
     NO_ACCOUNTS,
